+++ src/client/game/entities/item/weapons/EntityGuns.cpp
@@ -12,8 +12,8 @@ void EntityGlock::EventPickup(Character& picker_char)
 	m_Alive = false;
 }
 
-EntityGlock::EntityGlock(GameWorld *world, Entity *dropper, WeaponGlock *glock, const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_GLOCK, dropper, start_pos, Vec2d(12 * 4, 8 * 4))
+EntityGlock::EntityGlock(GameWorld *world, Entity *dropper, WeaponGlock *glock, const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_GLOCK, dropper, start_pos, Vec2f(12 * 4, 8 * 4))
 {
 	m_Glock = glock ? glock : new WeaponGlock(nullptr);
 }
@@ -30,8 +30,8 @@ void EntityShotgun::EventPickup(Character& picker_char)
 	m_Alive = false;
 }
 
-EntityShotgun::EntityShotgun(GameWorld *world, Entity *dropper, WeaponShotgun *shotgun, const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_SHOTGUN, dropper, start_pos, Vec2d(21 * 4, 6 * 4))
+EntityShotgun::EntityShotgun(GameWorld *world, Entity *dropper, WeaponShotgun *shotgun, const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_SHOTGUN, dropper, start_pos, Vec2f(21 * 4, 6 * 4))
 {
 	m_Shotgun = shotgun ? shotgun : new WeaponShotgun(nullptr);
 
@@ -48,8 +48,8 @@ void EntityBurst::EventPickup(Character& picker_char)
 	m_Burst = nullptr;
 }
 
-EntityBurst::EntityBurst(GameWorld *world, Entity *dropper, WeaponBurst *burst, const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_BURST, dropper, start_pos, Vec2d(20 * 4, 8 * 4))
+EntityBurst::EntityBurst(GameWorld *world, Entity *dropper, WeaponBurst *burst, const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_BURST, dropper, start_pos, Vec2f(20 * 4, 8 * 4))
 {
 	m_Burst = burst ? burst : new WeaponBurst(nullptr);
 
@@ -66,8 +66,8 @@ void EntitySniper::EventPickup(Character& picker_char)
 	m_Alive = false;
 }
 
-EntitySniper::EntitySniper(GameWorld *world, Entity *dropper, WeaponSniper *sniper, const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_SNIPER, dropper, start_pos, Vec2d(33 * 4, 9 * 4))
+EntitySniper::EntitySniper(GameWorld *world, Entity *dropper, WeaponSniper *sniper, const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_SNIPER, dropper, start_pos, Vec2f(33 * 4, 9 * 4))
 {
 	m_Sniper = sniper ? sniper : new WeaponSniper(nullptr);
 }
@@ -85,8 +85,8 @@ void EntityMinigun::EventPickup(Character& picker_char)
 	m_Minigun = nullptr;
 }
 
-EntityMinigun::EntityMinigun(GameWorld *world, Entity *dropper, WeaponMinigun *minigun, const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_MINIGUN, dropper, start_pos, Vec2d(17 * 4, 8 * 4))
+EntityMinigun::EntityMinigun(GameWorld *world, Entity *dropper, WeaponMinigun *minigun, const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_MINIGUN, dropper, start_pos, Vec2f(17 * 4, 8 * 4))
 {
 	m_Minigun = minigun ? minigun : new WeaponMinigun(nullptr);
 
@@ -106,8 +106,8 @@ void EntityPatersonNavy::EventPickup(Character& picker_char)
 EntityPatersonNavy::EntityPatersonNavy(GameWorld *world,
 									   Entity *dropper,
 									   PatersonNavy *paterson_navy,
-									   const Vec2d& start_pos)
-	: ItemEntity(world, ITEMTYPE_PATERSONNAVY, dropper, start_pos, Vec2d(72, 32))
+									   const Vec2f& start_pos)
+	: ItemEntity(world, ITEMTYPE_PATERSONNAVY, dropper, start_pos, Vec2f(72, 32))
 {
 	m_PatersonNavy = paterson_navy ? paterson_navy : new PatersonNavy(nullptr);
 }
