+++ src/client/game/weapons/projectile/PatersonNavy.h
@@ -10,12 +10,12 @@
 class PatersonNavy : public ProjectileWeapon
 {
 private:
-	double m_RecoilForce, m_BaseRecoilForce;
+	float m_RecoilForce, m_BaseRecoilForce;
 	double m_Damage;
 
 public:
-	static Vec2d sHoldPosition;
-	static std::pair<Vec2d, Vec2d> sHandPositions;
+	static Vec2f sHoldPosition;
+	static std::pair<Vec2f, Vec2f> sHandPositions;
 
 	explicit PatersonNavy(DirectionalEntity *parent);
