+++ src/client/game/weapons/projectile/ProjectileWeapon.h
@@ -42,24 +42,24 @@ protected:
 	WeaponType m_Type;
 	Texture *m_Texture;
 	Sound *m_ReloadSound;
-	Vec2d m_HoldPosition;
-	Vec2d m_LeftHandPosition;
-	Vec2d m_RightHandPosition;
+	Vec2f m_HoldPosition;
+	Vec2f m_LeftHandPosition;
+	Vec2f m_RightHandPosition;
 
 	// Ticking
 	virtual void TickTrigger();
 
 	// Generating
-	[[nodiscard]] double GenerateSpreadAngle() const;
-	[[nodiscard]] double GenerateRandomProjectileSpeed() const;
+	[[nodiscard]] float GenerateSpreadAngle() const;
+	[[nodiscard]] float GenerateRandomProjectileSpeed() const;
 
 public:
 	ProjectileWeapon(DirectionalEntity *parent,
 					 WeaponType type,
 					 Texture *texture,
 					 Sound *reload_sound,
-					 Vec2d *hold_position,
-					 std::pair<Vec2d, Vec2d> *hand_positions,
+					 Vec2f *hold_position,
+					 std::pair<Vec2f, Vec2f> *hand_positions,
 					 int tick_cooldown,
 					 int ammo_capacity,
 					 int total_ammo_capacity,
@@ -76,9 +76,9 @@ public:
 	[[nodiscard]] unsigned int AmmoCap() const { return m_AmmoCapacity; }
 	[[nodiscard]] unsigned long long TickCooldown() const { return m_TickCooldown; }
 	[[nodiscard]] unsigned long long LastShot() const { return m_LastShotAt; }
-	[[nodiscard]] Vec2d GetHoldPosition() const { return m_HoldPosition; }
-	[[nodiscard]] Vec2d GetLeftHand() const { return m_LeftHandPosition; }
-	[[nodiscard]] Vec2d GetRightHand() const { return m_RightHandPosition; }
+	[[nodiscard]] Vec2f GetHoldPosition() const { return m_HoldPosition; }
+	[[nodiscard]] Vec2f GetLeftHand() const { return m_LeftHandPosition; }
+	[[nodiscard]] Vec2f GetRightHand() const { return m_RightHandPosition; }
 
 	// Generating
 	[[nodiscard]] unsigned int NeededTrueAmmo() const;
