+++ src/client/game/weapons/projectile/WeaponBurst.h
@@ -18,8 +18,8 @@ private:
 	int m_BurstShotsLeft;
 
 public:
-	static Vec2d sHoldPosition;
-	static std::pair<Vec2d, Vec2d> sHandPositions;
+	static Vec2f sHoldPosition;
+	static std::pair<Vec2f, Vec2f> sHandPositions;
 
 	explicit WeaponBurst(Character *owner);
