+++ src/client/game/weapons/projectile/WeaponMinigun.h
@@ -20,8 +20,8 @@ private:
 	double m_Rotation;
 
 public:
-	static Vec2d sHoldPosition;
-	static std::pair<Vec2d, Vec2d> sHandPositions;
+	static Vec2f sHoldPosition;
+	static std::pair<Vec2f, Vec2f> sHandPositions;
 
 	explicit WeaponMinigun(Character *owner);
